feat:create upload icon feature
