+++ client/.env
@@ -1,0 +1,3 @@
+SUPABASE_PASSWORD=SyUbNjFraSH8iI4Z
+SUPABASE_KEY = eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImdhY25ncHNvZWticmlmeGFocGtnIiwicm9sZSI6ImFub24iLCJpYXQiOjE2OTM1Mzk4MTQsImV4cCI6MjAwOTExNTgxNH0.FsPo5sRvcobf8AchCc51g32FGRBVtuSw0oXSPKuJwms
+SUPABASE_URL = https://gacngpsoekbrifxahpkg.supabase.co
