+++ client/package.json
@@ -10,8 +10,12 @@
     "preview": "vite preview"
   },
   "dependencies": {
+<<<<<<< HEAD
     "@react-spring/web": "^9.7.3",
     "@material-tailwind/react": "^2.1.1",
+=======
+    "@supabase/supabase-js": "^2.33.1",
+>>>>>>> a94db27 (feat:create upload icon feature)
     "axios": "^1.5.0",
     "date-fns": "^2.30.0",
     "formik": "^2.4.3",
