+++ client/src/pages/AdminAddPackage.jsx
@@ -1,16 +1,13 @@
 import "../App.css";
 import AdminSideBar from "../components/AdminSideBar";
-import AdminNavBar from "../components/AdminNavBar";
-import { AdminAddPackageForm } from "../components/AdminAddPackageForm";
+// import AdminNavBar from "../components/AdminNavBar";
+import AdminAddPackageForm from "../components/AdminPackageForm";
 
 function AdminAddPackage() {
   return (
     <div className="flex flex-row h-screen">
       <AdminSideBar />
-      <div className="w-screen h-screen">
-        <AdminNavBar />
-        <AdminAddPackageForm />
-      </div>
+      <AdminAddPackageForm />
     </div>
   );
 }
