+++ server/apps/package.js
@@ -1,10 +1,17 @@
 import { Router } from "express";
 import { pool } from "../utils/db.js";
+import multer from "multer";
+import { cloudinaryUpload } from "../utils/upload.js";
 
 const packageRouter = Router();
 
+// const multerUpload = multer;
+// ({ dest: "uploads/" });
+
+// const avatarUpload = multerUpload.fields;
+// [{ name: "package_icon", maxCount: 2 }];
+
 packageRouter.post("/", async (req, res) => {
-  console.log(req.body);
   const {
     package_name,
     package_price,
@@ -12,6 +19,8 @@ packageRouter.post("/", async (req, res) => {
     package_icon,
     created_by,
   } = req.body;
+
+  // const packageIcon = await cloudinaryUpload(package_icon);
   try {
     await pool.query(
       `insert into packages (package_name,package_price,package_limit,package_icon,created_by,created_at,updated_at)
