+++ server/package.json
@@ -15,10 +15,12 @@
     "@supabase/supabase-js": "^2.33.1",
     "axios": "^1.5.0",
     "bcrypt": "^5.1.1",
+    "cloudinary": "^1.40.0",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2",
+    "multer": "^1.4.5-lts.1",
     "pg": "^8.11.3",
     "socket.io": "^4.7.2"
   },
